+++ spec/generators/pageflow/initializer/initializer_generator_spec.rb
@@ -5,9 +5,34 @@ require 'generators/pageflow/initializer/initializer_generator'
 module Pageflow
   module Generators
     describe InitializerGenerator, type: :generator do
-      it "generates 'config/initializers/pageflow.rb'" do
+      let(:initializer) { file('config/initializers/pageflow.rb') }
+
+      before do
         run_generator
-        expect(file('config/initializers/pageflow.rb')).to exist
+      end
+
+      it "generates 'config/initializers/pageflow.rb'" do
+        expect(initializer).to exist
+      end
+
+      it 'registers the built-in classic player controls' do
+        expect(initializer)
+          .to contain('config.widget_types.register(Pageflow::BuiltInWidgetType.classic_player_controls')
+      end
+
+      it 'registers the built-in slim player controls' do
+        expect(initializer)
+          .to contain('config.widget_types.register(Pageflow::BuiltInWidgetType.slim_player_controls')
+      end
+
+      it 'registers the built-in navigation widget type' do
+        expect(initializer)
+          .to contain('config.widget_types.register(Pageflow::BuiltInWidgetType.navigation')
+      end
+
+      it 'registers the built-in mobile_navigation widget type' do
+        expect(initializer)
+          .to contain('config.widget_types.register(Pageflow::BuiltInWidgetType.mobile_navigation')
       end
     end
   end