+++ CHANGELOG.md
@@ -4,6 +4,19 @@
 
 [Compare changes](https://github.com/codevise/pageflow/compare/12-0-stable...master)
 
+#### Breaking Changes
+
+- The built-in widget types must now be registered in the host application. To keep existing functionality from previous Pageflow versions, add these lines to `config/initializers/pageflow.rb` in your host application:
+
+```
+# Register the built-in widget types.
+# You can remove these or add different versions with the same name.
+config.widget_types.register(Pageflow::BuiltInWidgetType.navigation, default: true)
+config.widget_types.register(Pageflow::BuiltInWidgetType.mobile_navigation, default: true)
+config.widget_types.register(Pageflow::BuiltInWidgetType.classic_player_controls, default: true)
+config.widget_types.register(Pageflow::BuiltInWidgetType.slim_player_controls)
+```
+
 Media stack:
 
 - Switch from `Expires` to `Cache-Control` header for media uploads.